persist: split Batch and BatchBuilder out into a module
The impl is already decently long for them and it's about to get longer.
